requirements.txt
imutils==0.5.4
Jinja2==3.1.1
joblib==1.1.0
MarkupSafe==2.1.1
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.0
PyMySQL==1.0.2
pyparsing==3.0.8
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyqtdarktheme==1.1.0
PyQtWebEngine==5.15.5
PyQtWebEngine-Qt5==5.15.2
pyserial==3.5
PySide6==6.3.0
PySide6-Addons==6.3.0
PySide6-Essentials==6.3.0
pytesseract==0.3.9
qt-material==2.10
scikit-learn==1.0.2
scipy==1.8.0
shiboken6==6.3.0
threadpoolctl==3.1.0
